+++ lib/td_dd/loader.ex
@@ -17,6 +17,10 @@ defmodule TdDd.Loader do
 
   require Logger
 
+  @typep multi_error :: {:error, Multi.name(), any(), %{required(Multi.name()) => any()}}
+  @typep multi_success :: {:ok, map()}
+  @typep multi_result :: multi_success() | multi_error()
+
   def load(records, audit, opts \\ [])
 
   def load(%{structures: structure_records, fields: field_records} = records, audit, opts) do
@@ -48,8 +52,15 @@ defmodule TdDd.Loader do
     multi(structure_records, relation_records, audit, opts)
   end
 
-  @spec multi([map], [map], %{ts: DateTime.t()}, Keyword.t()) ::
-          {:ok, map} | {:error, Multi.name(), any(), %{required(Multi.name()) => any()}}
+  @spec replace_mutable_metadata([map], TdDd.Systems.System.t(), map, Keyword.t()) ::
+          multi_result()
+  def replace_mutable_metadata(records, system, audit, opts) do
+    count = Enum.count(records)
+    Logger.info("Starting mutable metadata bulk replace (#{count} records)")
+    metadata_multi(records, system, audit, opts[:operation])
+  end
+
+  @spec multi([map], [map], %{ts: DateTime.t()}, Keyword.t()) :: multi_result()
   def multi(structure_records, relation_records, %{ts: ts} = audit, opts \\ []) do
     Multi.new()
     |> Multi.run(:graph, LoadGraph, :load_graph, [structure_records, relation_records, opts])
@@ -66,14 +77,32 @@ defmodule TdDd.Loader do
       opts[:source],
       ts
     ])
-    |> Multi.run(:delete_metadata, Metadata, :delete_missing_metadata, [ts])
-    |> Multi.run(:replace_metadata, Metadata, :replace_metadata, [structure_records, ts])
+    |> replace_or_merge_metadata(structure_records, ts, opts[:operation])
     |> Multi.run(:structure_ids, __MODULE__, :structure_ids, [])
     |> Multi.run(:system_ids, fn _, _ -> {:ok, system_ids(structure_records)} end)
     |> Multi.run(:classification, fn _, %{system_ids: ids} ->
       Classifiers.classify_many(ids, updated_at: ts)
     end)
     |> Repo.transaction()
+  end
+
+  @spec metadata_multi([map], TdDd.Systems.System.t(), map, binary()) :: multi_result()
+  def metadata_multi(records, system, %{ts: ts} = _audit, operation \\ "replace") do
+    Multi.new()
+    |> Multi.run(:missing_external_ids, Metadata, :missing_external_ids, [records, system])
+    |> replace_or_merge_metadata(records, ts, operation)
+    |> Multi.run(:structure_ids, __MODULE__, :structure_ids, [])
+    |> Repo.transaction()
+  end
+
+  defp replace_or_merge_metadata(multi, structure_records, ts, "merge") do
+    Multi.run(multi, :merge_metadata, Metadata, :merge_metadata, [structure_records, ts])
+  end
+
+  defp replace_or_merge_metadata(multi, structure_records, ts, _replace) do
+    multi
+    |> Multi.run(:delete_metadata, Metadata, :delete_missing_metadata, [ts])
+    |> Multi.run(:replace_metadata, Metadata, :replace_metadata, [structure_records, ts])
   end
 
   def system_ids(structure_records) do
@@ -98,7 +127,9 @@ defmodule TdDd.Loader do
   defp structure_ids({:update_source_ids, {_, ids}}), do: ids
   defp structure_ids({:delete_metadata, {_, ids}}), do: ids
   defp structure_ids({:replace_metadata, ids}), do: ids
+  defp structure_ids({:merge_metadata, ids}), do: ids
   defp structure_ids({:insert_versions, {_, dsvs}}), do: Enum.map(dsvs, & &1.data_structure_id)
   defp structure_ids({:update_versions, {_, dsvs}}), do: Enum.map(dsvs, & &1.data_structure_id)
   defp structure_ids({:replace_versions, {_, dsvs}}), do: Enum.map(dsvs, & &1.data_structure_id)
+  defp structure_ids({:missing_external_ids, _}), do: []
 end